Add feature to change number of squares per side
Button, when pressed, prompts user to enter number of squares.
Default is 16.

Problem: can't access the local variable "squares" outside the scope of
the event listener block. So creating a grid inside the that block
makes it so that the trailing feature doesn't take effect unless it's
also re-declared inside the event listener block (redundant code).

Other things:
- Refactored creating grid into a function
- Create removeGrid function